+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Prometheus rules availability plugin.
 - Prometheus targets availability plugin.
 - Kooper handling availability plugin.
 - Kooper handling latency plugin.